duplicate_detector/requirements.txt
# in requirements.txt
groq
firecrawl-py
google-api-python-client
python-dotenv
nltk
scikit-learn
